# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/artist": {
            "get": {
                "tags": [
                    "Artists"
                ],
                "summary": "List all artist",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Artists"
                ],
                "summary": "Create a new artist",
                "parameters": [
                    {
                        "description": "Artist details",
                        "name": "artist",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Artist"
                        }
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "example": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Artist Created Successfully",
                        "schema": {
                            "$ref": "#/definitions/models.Artist"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Artists"
                ],
                "summary": "Delete a specific artist",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID of the artist",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Artists"
                ],
                "summary": "Update a specific artist",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID of the artist",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Artist details",
                        "name": "artist",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Artist"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Artist updated successfully",
                        "schema": {
                            "$ref": "#/definitions/models.Artist"
                        }
                    }
                }
            }
        },
        "/artist/name": {
            "get": {
                "tags": [
                    "Artists"
                ],
                "summary": "Search artist by name",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"Emicida\"",
                        "description": "Name of the artist",
                        "name": "name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/artist/specific": {
            "get": {
                "tags": [
                    "Artists"
                ],
                "summary": "List an artist by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID of the artist",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/ping": {
            "get": {
                "description": "do ping",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Test API connection"
                ],
                "summary": "Test API connection",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Artist": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "album": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "released": {
                    "type": "string"
                }
            }
        }
    }
}